Move useCouncilSize hook to council
